I am playing the Game Fallout 4, but I am having some Modding Issues.
Currently, a Mod called "Scrap Everything" quite working. I have tried uninstalling it 4 different ways available.  I have been using it for over 6 months without issue, but now it is saying it is missing master files. but if i go to plugins, the files are there there, but others not mentioned are missing? so confusing. it says: "This mod does not have any known dependencies other than the base game." "Put this file at the absolute END of your load order, or else you will get areas in which scrapping doesn't work as advertised!" which i have! yet the Data files show the missing files.. I'm at a loss on how to fix the not working issue.. can anyone help.